feat: invoke websoc proxy as code outside prod
