+++ libs/websoc-utils/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "@libs/websoc-utils",
+  "version": "0.0.0",
+  "private": true,
+  "type": "module",
+  "main": "src/index.ts",
+  "types": "src/index.ts"
+}
