+++ tools/cdk/package.json
@@ -10,7 +10,6 @@
   },
   "dependencies": {
     "dotenv": "16.3.1",
-    "websoc-proxy-service": "*",
     "websoc-scraper-v2": "*"
   },
   "devDependencies": {
